rules-dsl/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
antlr4-python3-runtime==4.13.2
PyYAML==6.0.1
